+++ @twc_config/addons/twc_millennial/config.cpp
@@ -1,176 +1,271 @@
 class CfgPatches
 {
-	class twc_1990
-	{
-		units[]=
-		{
-			"1990_British_Base",
-			"1990_British_SectionCommander",
-			"1990_British_Rifleman",
-			"1990_British_Gunner",
-			"1990_British_2IC",
-			"1990_British_CGG",
-			"1990_British_CGAss",
-			"1990_British_Platoon_Commander",
-			"1990_British_Platoon_Sergeant",
-			"1990_British_Medic",
-			"1990_British_HeliPilot",
-			"1990_British_JetPilot",
-			"1990_British_MilanGunner",
-			"1990_British_MilanAssistant",
-			"1990_British_MortarGunner",
-			"1990_British_FAC",
-			"1990_British_Rifleman_Mechanised",
-			"1990_British_Gunner_Mechanised",
-			"1990_British_2IC_Mechanised",
-			"1990_British_Vehicle_Commander",
-			"1990_British_Vehicle_Crew",
-			
-			"1990_British_SectionCommander_Desert",
-			"1990_British_Rifleman_Desert",
-			"1990_British_CGG_Desert",
-			"1990_British_CGAss_Desert",
-			"1990_British_Gunner_Desert",
-			"1990_British_2IC_Desert",
-			"1990_British_Platoon_Commander_Desert",
-			"1990_British_Platoon_Sergeant_Desert",
-			"1990_British_Medic_Desert",
-			"1990_British_HeliPilot_Desert",
-			"1990_British_MilanGunner_Desert",
-			"1990_British_MilanAssistant_Desert",
-			"1990_British_MortarGunner_Desert",
-			"1990_British_FAC_Desert",
-			"1990_British_Rifleman_Mechanised_Desert",
-			"1990_British_Gunner_Mechanised_Desert",
-			"1990_British_2IC_Mechanised_Desert",
-			"1990_British_Vehicle_Commander_Desert",
-			"1990_British_Vehicle_Crew_Desert",
-			"1990_British_Tank_Commander_Desert",
-			"1990_British_Tank_Crew_Desert"
-			
-/*			
-			"2000_British_Base",
-			"2000_British_SectionCommander",
-			"2000_British_Rifleman",
-			"2000_British_Rifleman_AT",
-			"2000_British_AutoRifleman",
-			"2000_British_2IC",
-			"2000_British_Marksman",
-			"2000_British_Platoon_Commander",
-			"2000_British_Platoon_Sergeant",
-			"2000_British_Platoon_CSM",
-			"2000_British_Medic",
-			"2000_British_MilanGunner",
-			"2000_British_MilanAssistant",
-			"2000_British_MortarGunner",
-			"2000_British_FAC",
-			"2000_British_CGG",
-			"2000_British_CGAss",
-			"2000_British_Rifleman_Mounted",
-			"2000_British_Rifleman_AT_Mounted",
-			"2000_British_AutoRifleman_Mounted",
-			"2000_British_2IC_Mounted",
-			"2000_British_Marksman_Mounted",
-			"2000_British_Vehicle_Commander",
-			"2000_British_Vehicle_Crew",
-			
-			"2000_British_SectionCommander_Desert",
-			"2000_British_Rifleman_Desert",
-			"2000_British_Rifleman_AT_Desert",
-			"2000_British_AutoRifleman_Desert",
-			"2000_British_2IC_Desert",
-			"2000_British_Marksman_Desert",
-			"2000_British_Platoon_Commander_Desert",
-			"2000_British_Platoon_Sergeant_Desert",
-			"2000_British_Platoon_CSM_Desert",
-			"2000_British_Medic_Desert",
-			"2000_British_MilanGunner_Desert",
-			"2000_British_MilanAssistant_Desert",
-			"2000_British_MortarGunner_Desert",
-			"2000_British_FAC_Desert",
-			"2000_British_CGG_Desert",
-			"2000_British_CGAss_Desert",
-			"2000_British_Rifleman_Mounted_Desert",
-			"2000_British_Rifleman_AT_Mounted_Desert",
-			"2000_British_Marksman_Mounted_Desert",
-			"2000_British_AutoRifleman_Mounted_Desert",
-			"2000_British_2IC_Mounted_Desert",
-			"2000_British_Vehicle_Commander_Desert",
-			"2000_British_Vehicle_Crew_Desert",
-			"2000_British_Tank_Commander_Desert",
-			"2000_British_Tank_Crew_Desert"
-*/
-		};
-		weapons[]={};
-		requiredVersion=0.1;
-		requiredAddons[]=
-		{
+	class twc_millennial {
+		units[] = {};
+		weapons[] = {};
+		requiredVersion = 0.1;
+		requiredAddons[] = {
 			"A3_Characters_F_BLUFOR",
-			"TWC_Faction"
+			"A3_Characters_F_OPFOR",
+			"twc_faction",
+			"twc_ai"
 		};
 	};
-};
-class EventHandlers;
-class CfgEditorSubCategories
-{
-	class Men_1990_Arctic_Dismounted
-	{
-		displayName = "Men (90's Arctic Dismounted)";
-	};
-	class Men_1990_Desert_Dismounted
-	{
-		displayName = "Men (90's Desert Dismounted)";
-	};
-	class Men_1990_Desert_Mounted
-	{
-		displayName = "Men (90's Desert Mounted)";
-	};
-	class Men_1990_Woodland_Dismounted
-	{
-		displayName = "Men (90's Woodland Dismounted)";
-	};
-	class Men_1990_Woodland_Mounted
-	{
-		displayName = "Men (90's Woodland Mounted)";
-	};
-/*
-	class Men_2000_Arctic_Dismounted
-	{
-		displayName = "Men (00's Arctic Dismounted)";
-	};
-	class Men_2000_Desert_Dismounted
-	{
-		displayName = "Men (00's Desert Dismounted)";
-	};
-	class Men_2000_Desert_Mounted
-	{
-		displayName = "Men (00's Desert Mounted)";
-	};
-	class Men_2000_Woodland_Dismounted
-	{
-		displayName = "Men (00's Woodland Dismounted)";
-	};
-	class Men_2000_Woodland_Mounted
-	{
-		displayName = "Men (00's Woodland Mounted)";
-	};
-*/
 };
 
-#include "compositions.hpp"
+class EventHandlers;
+
+class CfgWeapons
+{
+	#include "weapons.hpp"
+};
 
 class CfgVehicles
 {
-	// Time saving MACROs - NC = Non-Conflict
-	#define maga_nc(a,b) class _nc_##a {magazine = a; count = b;}
-	#define weap_nc(a,b) class _nc_##a {weapon = a; count = b;}
-	#define item_nc(a,b) class _nc_##a {name = a; count = b;}
-	
-	//Needed for 1990 + 2000 base class
-	class B_Soldier_base_F;
-	
-	#include "1990.hpp"	
-//	#include "2000.hpp"	
 	#include "backpacks.hpp"
+
+	class B_Soldier_base_F;
+	class O_Soldier_base_F;
+	class I_Soldier_F;
+
+	#define MEDICAL_LOADOUT "ACE_EarPlugs","ACE_fieldDressing","ACE_fieldDressing","ACE_fieldDressing","ACE_fieldDressing","ACE_packingBandage","ACE_packingBandage","ACE_packingBandage","ACE_packingBandage","ACE_packingBandage","ACE_packingBandage","ACE_elasticBandage","ACE_elasticBandage","ACE_elasticBandage","ACE_elasticBandage","ACE_elasticBandage","ACE_elasticBandage","ACE_quikclot","ACE_quikclot","ACE_quikclot","ACE_quikclot","ACE_morphine","ACE_morphine","ACE_epinephrine","ACE_tourniquet","ACE_tourniquet"
+	#define MAG_2(a) a, a
+	#define MAG_3(a) a, a, a
+	#define MAG_4(a) a, a, a, a
+	#define MAG_5(a) a, a, a, a, a
+	#define MAG_6(a) a, a, a, a, a, a
+	#define MAG_7(a) a, a, a, a, a, a, a
+	#define MAG_8(a) a, a, a, a, a, a, a, a
+	#define MAG_9(a) a, a, a, a, a, a, a, a, a
+	#define MAG_10(a) a, a, a, a, a, a, a, a, a, a
+	#define MAG_11(a) a, a, a, a, a, a, a, a, a, a, a
+	#define MAG_12(a) a, a, a, a, a, a, a, a, a, a, a, a
+	#define CATEGORY(a) editorSubcategory = a;\
+			vehicleClass = a;
+
+	//GENERIC
+	class TWC_Infantry_Millennial_Base: B_Soldier_base_F
+	{
+		scope = 1;
+		displayName = "Base";
+		faction = "TWC_Millennial";
+		CATEGORY(TWC_Infantry_Regular_Woodland)
+		icon = "iconMan";
+		nakedUniform = "U_BasicBody";
+		uniformClass = "UK3CB_BAF_U_Smock_DPMW";
+		backpack = "";
+		linkedItems[] =
+		{
+			"UK3CB_BAF_V_PLCE_Webbing_DPMW",
+			"UK3CB_BAF_H_Mk6_DPMW_A",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		respawnLinkedItems[] =
+		{
+			"UK3CB_BAF_V_PLCE_Webbing_DPMW",
+			"UK3CB_BAF_H_Mk6_DPMW_A",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		Items[] =
+		{
+			MEDICAL_LOADOUT
+		};
+		respawnItems[] =
+		{
+			MEDICAL_LOADOUT
+		};
+		weapons[] =
+		{
+			"Throw",
+			"Put"
+		};
+		respawnweapons[] =
+		{
+			"Throw",
+			"Put"
+		};
+		magazines[] = {};
+		respawnmagazines[] = {};
+	};
+	class TWC_Infantry_Millennial_Plane_Pilot: TWC_Infantry_Millennial_Base
+	{
+		scope = 2;
+		displayName = "Jet Pilot";
+		faction = "TWC_Millennial";
+		CATEGORY(TWC_ForceType_Airborne)
+		uniformClass = "UK3CB_BAF_U_HeliPilotCoveralls_RAF";
+		linkedItems[] =
+		{
+			"CUP_V_B_PilotVest",
+			"UK3CB_BAF_H_PilotHelmetHeli_A",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		respawnLinkedItems[] =
+		{
+			"CUP_V_B_PilotVest",
+			"UK3CB_BAF_H_PilotHelmetHeli_A",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		Items[] =
+		{
+			MEDICAL_LOADOUT,
+			"ACE_MapTools"
+		};
+		respawnItems[] =
+		{
+			MEDICAL_LOADOUT,
+			"ACE_MapTools"
+		};
+		weapons[] =
+		{
+			"Throw",
+			"Put"
+		};
+		respawnweapons[] =
+		{
+			"Throw",
+			"Put"
+		};
+		magazines[] =
+		{
+			"SmokeShell",
+			"SmokeShell"
+		};
+		respawnmagazines[] =
+		{
+			"SmokeShell",
+			"SmokeShell"
+		};
+	};
+	class TWC_Infantry_Millennial_Tank_Commander: TWC_Infantry_Millennial_Base
+	{
+		scope = 2;
+		displayName = "Tank Commander";
+		faction = "TWC_Millennial";
+		CATEGORY(TWC_ForceType_Armoured)
+		uniformClass = "UK3CB_BAF_U_CrewmanCoveralls_RTR";
+		linkedItems[] =
+		{
+			"SP_P58_UrbanPatrol",
+			"UK3CB_BAF_H_Beret_RTR_PRR",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		respawnLinkedItems[] =
+		{
+			"SP_P58_UrbanPatrol",
+			"UK3CB_BAF_H_Beret_RTR_PRR",
+			"ItemCompass",
+			"itemMap",
+			"ItemWatch"
+		};
+		Items[] =
+		{
+			MEDICAL_LOADOUT,
+			"ACE_MapTools"
+		};
+		respawnItems[] =
+		{
+			MEDICAL_LOADOUT,
+			"ACE_MapTools"
+		};
+		weapons[] =
+		{
+			"TWC_Weapon_L85A1",
+			"Throw",
+			"Put"
+		};
+		respawnweapons[] =
+		{
+			"TWC_Weapon_L85A1",
+			"Throw",
+			"Put"
+		};
+		magazines[] =
+		{
+			MAG_3("UK3CB_BAF_556_30Rnd"),
+			MAG_2("SmokeShell")
+		};
+		respawnmagazines[] =
+		{
+			MAG_3("UK3CB_BAF_556_30Rnd"),
+			MAG_2("SmokeShell")
+		};
+	};
+	class TWC_Infantry_Millennial_Tank_Crew: TWC_Infantry_Millennial_Tank_Commander
+	{
+		displayName = "Tank Crew";
+	};
+	#include "1990_regular\woodland.hpp"
+	#include "1990_regular\woodland_dismounted.hpp"
+	#include "1990_regular\desert.hpp"
+	#include "1990_regular\desert_dismounted.hpp"
+
+	#include "1990_coin\woodland.hpp"
+	#include "1990_coin\woodland_dismounted.hpp"
+	#include "1990_coin\desert.hpp"
+	#include "1990_coin\desert_dismounted.hpp"
+
+	/*#include "2000_regular\woodland.hpp"
+	#include "2000_regular\woodland_dismounted.hpp"
+	#include "2000_regular\desert.hpp"
+	#include "2000_regular\desert_dismounted.hpp"
+
+	#include "2000_coin\woodland.hpp"
+	#include "2000_coin\woodland_dismounted.hpp"
+	#include "2000_coin\desert.hpp"
+	#include "2000_coin\desert_dismounted.hpp"*/
 };
-#include "weapons.hpp"
+
+class CfgGroups
+{
+	class WEST
+	{
+		class TWC_Groups_Millennial
+		{
+			name = "TWC Groups Millennial";
+			#include "1990_regular\groups.hpp"
+			#include "1990_coin\groups.hpp"
+			class TWC_Groups_Millennial_Generic
+			{
+				name = "Generic";
+				class Tank_Crew
+				{
+					name = "Tank Crew";
+					faction = "TWC_Millennial";
+					side = 1;
+					class Unit0
+					{
+							side = 1;
+							vehicle = "TWC_Infantry_Millennial_Tank_Commander";
+							rank = "SERGEANT";
+							position[] = {0,0,0};
+					};
+					class Unit1
+					{
+							side = 1;
+							vehicle = "TWC_Infantry_Millennial_Tank_Crew";
+							rank = "PRIVATE";
+							position[] = {2,0,0};
+					};
+					class Unit2
+					{
+							side = 1;
+							vehicle = "TWC_Infantry_Millennial_Tank_Crew";
+							rank = "PRIVATE";
+							position[] = {4,0,0};
+					};
+				};
+			};
+		};
+	};
+};
