--- @twc_config/addons/twc_millennial_new/backpacks.hpp
@@ -1,28 +1,0 @@
-#define ADD_MAGA(a,b) class _nc_##a {magazine = a; count = b;}
-#define ADD_WEAP(a,b) class _nc_##a {weapon = a; count = b;}
-#define ADD_ITEM(a,b) class _nc_##a {name = a; count = b;}
-
-class SP_Backpack_BlowpipeCarrier;
-
-//Woodland
-class UK3CB_BAF_B_Bergen_DPMW_SL_A;
-class UK3CB_BAF_B_Bergen_DPMW_Rifleman_A;
-class UK3CB_BAF_B_Bergen_DPMW_Rifleman_B;
-class UK3CB_BAF_B_Bergen_DPMW_JTAC_H_A;
-
-//Desert
-class UK3CB_BAF_B_Bergen_DDPM_SL_A;
-class UK3CB_BAF_B_Bergen_DDPM_Rifleman_A;
-class UK3CB_BAF_B_Bergen_DDPM_Rifleman_B;
-class UK3CB_BAF_B_Bergen_DDPM_JTAC_H_A;
-
-//Arctic
-class UK3CB_BAF_B_Bergen_Arctic_SL_A;
-class UK3CB_BAF_B_Bergen_Arctic_Rifleman_A;
-class UK3CB_BAF_B_Bergen_Arctic_Rifleman_B;
-class UK3CB_BAF_B_Bergen_Arctic_JTAC_H_A;
-
-#include "1990_coin\backpacks.hpp"
-#include "1990_regular\backpacks.hpp"
-#include "2000_coin\backpacks.hpp"
-#include "2000_regular\backpacks.hpp"
